pip-requirements.txt
Package             Version     Build
------------------- ----------- -----
python              3.12.10
cffi                2.0.0
click               8.3.0
click-default-group 1.2.4
clr_loader          0.2.8
colorama            0.4.6
cryptography        46.0.3
numpy               2.1.3
opencv-python       4.12.0.88
openvino            2024.6.0    17404
openvino-telemetry  2025.2.0
packaging           25.0
pillow              12.0.0
pip                 25.3
pluggy              1.6.0
psycopg2-binary     2.9.11
pycparser           2.23
PyQt6               6.10.0
PyQt6-Qt6           6.10.0
PyQt6_sip           13.10.2
python-dateutil     2.9.0.post0
python-dotenv       1.2.1
pythonnet           3.0.5
pyzkfp              0.1.5
scipy               1.16.2
six                 1.17.0
sqlite-fts4         1.0.3
sqlite-utils        3.38
tabulate            0.9.0